refactor: Update execution time metrics to use double for precision and clean up code formatting

+++ Apps/Server/CairoTransportation/Utils/Helpers/Common/DTOs/AlgorithmTraceDto.cs
@@ -4,6 +4,6 @@ public class AlgorithmTraceDto
 {
     public int VisitedNodes { get; set; }
     public int ExpandedNodes { get; set; }
-    public long ExecutionTimeMs { get; set; }
+    public double ExecutionTimeMs { get; set; }
 }
